CaseNote, PersonNote, CategoryForCase, ConvictionType should have composite keys
-CaseNote, PersonNote: I think composite keys are better here because
CaseNote/PersonNote represent weak entities, and so they are only
determined by the key (caseNoteId, caseId). This change lead to
dropping the auto_increment on caseNoteId. The system said that
auto_increment is only supported if the attribute is the sole primary
key.
-Also, I specified a composite key for CategoryForCase. As the
relationship is N:M both attributes are part of the key.
-Similarly ConvictionType is N:M and should have a composite key
-I think when a Category gets deleted, the ConvictionTypes pointing to
it should be deleted as well instead having their references being set
to null - there is no value in a ConvictionType tuple with
null-categoryName
-I suggest the following syntactical convection: whenever possible,
write the constraints on the same line as the attribute, otherwise
write it below the attributes.