-More enforced foreign key constraints -Add: test set of illegal inserts
-Convicted and CaseNote still had some not enforced foreign key
constraints.
-dataTestSetInsertsThatShouldBeRejected.txt is a set of queries that
should be rejected on a fresh database by the system because of foreign
key constraints. It can be used to test if the system/model enforces
the constraints. If the dataTestSet is already loaded, the queries
should be rejected because of duplicate primary keys.
-The model should be pretty ok now. If you don’t have any additional
comments, I think we can freeze the model.