PersonCreation
updated:
PersonCreation.jsp
Header.jsp
   — updated the form to create person. I don’t know how to handle the
date. It seems that we need html5 to have the type form « date ». We
can either set Eclipse for html5, create a long form of type « select »
or just leave it like this and check if it’s a valid date with java.
(My preference is for the last option, because create manually select
form is kind of ugly style… But we can still do that if you want!)

NB : I will try do focus on this project in the beginning of the week,
like that I’ve all the end of the week and the weekend for the other.
But, if you have comments, asks (i.e. , stuff you want I do) , feel
free to say to me : I’m not delicate. (Not clever enough for that! :-) )