Add CaseNote implemented
Case.jsp, CaseServlet
(/=) send the caseId as part of the request, and not as a session
attribute. This streamlines the CaseServlet a bit.

CaseDatastore
(+) addCaseNote, implemented using a lock over the class, a query to
get the next id and a query to do the insertion. It would be nicer to
do it in a single query, but there is a problem with using the same
table in the from clause of a nested query in an insertion.