Removed ConvictionType table, adapt queries
(/=) I had several issues with the implementation of the new search
methods. The main problem was with the datastore interface: When a
specific date is wanted and used for comparison, as in the between
queries the datastore should take already parsed dates. The datastore
should not rely on its clients to call its methods with properly
formatted date string.

(/=) this change required to modify the search servlet.

(/=) birthdates weren’t working so I fixed it (bug in Servlet & bug in
datastore)

(/=) An unrelated issue was that I wasn’t able to import the library
needed for Pair. Since its a tiny class, I just reimplemented it.

(/=) I have modified all of the datastore interface methods to take
util.Date, as this is more general, and the interpretation should
remain clear.