(/=) SQLLexer : Consider the fact that SQL syntax is case insensitive
This is accounted for by passing the case insensitive flag (?i) to the
Pattern. We can write the rules just accounting for a-z now, and FROM,
WHERE, … now also tokenize correctly.
Additionally, the matches (except literals) are converted to lowercase
for later convenience (semantic is case insensitive, so having it all
in lowercase is handy).