(+) SQLGrammar first draft (/=) SQLToken fixes
SQLToken:
Case.* and Case.name are now treated as different lexicographical
classes. They cannot be used in the same places, so this distinction
helps specify the Grammar.

(+) SQLGrammar:
Initial specification of insert and select.
For now, the grammar is in no normal form, as the algorithm to use for
parsing is not fixed yet. For now, the grammar uses lambda-productions
for optional clauses, but for an implementation it might be more
convenient to eliminate them.