(+) Implement the basic parser for insert statements
I chose to implement the parser manually, based directly on the grammar
(top down recursive). This is easy now, as from the form of the grammar
no backtracking is necessary. This also means the method takes linear
time in the length of the token sequence.
(+) Right now, no tree is constructed, but this can be changed later by
adding appropriate actions.
(+) A simple test class makes sure two canonical examples succeed and
some examples fail.