Creation of a tree for deletion
I’ve chosen to not create a new class for the operators, since it
sounds very duplicate and, even for predicate, seems to me not
necessary.

Updated:
-Predicate
Now, predicate return an ASTNode of the form:
where—————————————
|                 |                 |                   |
a               ==               b                  optional AND

Please not advance before giving to predicate. (The position of the
TokenStream should be at the WHERE or AND token)

-Delete
Create the tree for delete with second child the where-clause and first
child the table of concern.