Tree for insert
Updated :
The insert parsing should now create a tree. The
DescriptionSubTrees.txt has been modified accordingly.

It’s my last commit of the day, I will continue tomorrow morning. (I
switch to the other project and… Actually, I discovered my first
allergy: manioc. >.<) I needed some time to get everything, but now it
seems to be clear to me what still has to be done. I hope we are really
close to the end of the parsing! I even think that for update, delete
and insert we don’t need any rewriting and can use the trees as they
are.

My todo list for tomorrow :
check and test! And go on.

Have a nice day!