Schema Interface
Added : TableSchema.java

This is the first proposition for the Schema of a table. It is meant to
be accessed through the connection (as every DB request should be).
Once the table schema is created, every client can use it in order to
get the keys names, attributes names and types, etc.

I think the implementation of this class should be rather simple
because it’s just an access to the metadata of the table.

Do you need other methods for the schema?