(/=) Updated SQLOperator interface
Now make use of TableSchema, and there’s a better mechanism for
inheritance using a varargs children parameter (similar to how the
SyntaxTreeNodes are implemented).
The Operators do not perform any type inference (anymore), as this has
already been done on the syntax tree. Operators just assume their
schema to be correct and consistent with their children.